requirements.txt
anyio==4.4.0
asyncping3==3.1.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
pillow==10.4.0
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
texttable==1.7.0
urllib3==2.2.2
